added chain and manticore to the map list for gamemode, increased starting points per match to account for bots being killed removing points, set custom bot count for each map, fixed cap points not properly networking the points taking them gives, bots in a squad now copy their leader's path rather than trying to walk to their leader's position.

+++ sb/entities/entities/daktank_cap/init.lua
@@ -98,7 +98,7 @@ function ENT:Think()
 				end
 				net.Start( "DT_killnotification" )
 					net.WriteInt(3, 32)
-					net.WriteInt(PointsGained, 32)
+					net.WriteFloat(PointsGained, 32)
 				net.Send( v )
 				if not(v:InVehicle()) then v:addPoints( PointsGained ) end
 			end
@@ -122,7 +122,7 @@ function ENT:Think()
 				end
 				net.Start( "DT_killnotification" )
 					net.WriteInt(3, 32)
-					net.WriteInt(PointsGained, 32)
+					net.WriteFloat(PointsGained, 32)
 				net.Send( v )
 				if not(v:InVehicle()) then v:addPoints( PointsGained ) end
 			end
